Deprecated `pending_mentions` command in favour of `mentions_pending`.
Naming consistency with `mentions_reverify` and any future commands.
